Refactored code and fixed song length data type

+++ app/src/main/java/com/example/songbytesapp/Song.java
@@ -2,10 +2,10 @@ package com.example.songbytesapp;
 
 public class Song {
     private String name;
-    private int length;
+    private double length;
     private String artist;
 
-    public Song(String name, int length, String artist) {
+    public Song(String name, double length, String artist) {
         this.name = name;
         this.length = length;
         this.artist = artist;
@@ -19,11 +19,11 @@ public class Song {
         this.name = name;
     }
 
-    public int getLength() {
+    public double getLength() {
         return length;
     }
 
-    public void setLength(int length) {
+    public void setLength(double length) {
         this.length = length;
     }
